More changes to first level
It's pretty playable now. The timer has been implemented, the enemies
can move, and you can enter the classroom.

+++ Assets/Test Stuff/InventoryManager.js
@@ -1,6 +1,6 @@
 ﻿#pragma strict
 
 
-static var pencil : int = 0;
-static var essay : int = 0;
-static var secret : int = 0;
+static var pencil : float = 0;
+static var essay : float = 0;
+static var secret : float = 0;
